Fixed bug on accelereometer data conversion, now we can rad and print correct value of 3-Axis acceleration. Need to covert accelerations in G form

+++ VirtualSense/DJ_VirtualMachine.1.0/darjeeling.1.1/src/lib/virtualsense/javax/virtualsense/sensors/evm/accelerometer/AccVector.java
@@ -28,7 +28,7 @@ package javax.virtualsense.sensors.evm.accelerometer ;
  */
 public class AccVector
 {	
-	public int x;
-	public int y;
-	public int z;
+	public short x;
+	public short y;
+	public short z;
 }
